feat(logic): move game logic files into `logic/` folder.
